# internal/http/openapi_spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GoClaw Gateway API",
    "description": "PostgreSQL multi-tenant AI agent gateway with WebSocket RPC + HTTP API.\n\n## Authentication\n\nAll endpoints require a Bearer token in the `Authorization` header:\n\n```\nAuthorization: Bearer <gateway-token-or-api-key>\n```\n\nYou can use either the **gateway token** (grants admin access) or an **API key** created via the API Keys endpoints (grants scoped access).\n\nIf no token is configured on the server, authentication is disabled (backward compatibility).\n\n## Common Headers\n\n| Header | Description |\n|--------|-------------|\n| `X-ArgoClaw-User-Id` | External user ID for multi-tenant context |\n| `X-ArgoClaw-Agent-Id` | Target agent ID (alternative to model prefix) |\n| `Accept-Language` | Locale for error messages (`en`, `vi`, `zh`) |\n\n## WebSocket Protocol\n\nConnect via `POST /ws` (upgrade). Protocol v3 uses frame types: `req`, `res`, `event`.\nFirst request must be `connect` with `{\"token\": \"...\", \"user_id\": \"...\", \"locale\": \"en\"}`.",
    "version": "0.2.0",
    "contact": {
      "name": "GoClaw",
      "url": "https://github.com/nextlevelbuilder/goclaw"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/providers": {
      "get": {
        "tags": ["Providers"],
        "summary": "List LLM providers",
        "responses": { "200": { "description": "Provider list (API keys masked)" } }
      },
      "post": {
        "tags": ["Providers"],
        "summary": "Create LLM provider",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProviderInput" } } } },
        "responses": { "201": { "description": "Provider created" } }
      }
    },
    "/v1/providers/{id}": {
      "get": {
        "tags": ["Providers"],
        "summary": "Get provider",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Provider details (API key masked)" } }
      },
      "put": {
        "tags": ["Providers"],
        "summary": "Update provider",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Provider updated" } }
      },
      "delete": {
        "tags": ["Providers"],
        "summary": "Delete provider",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Provider deleted" } }
      }
    },
    "/v1/providers/{id}/models": {
      "get": {
        "tags": ["Providers"],
        "summary": "List available models",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Model list" } }
      }
    },
    "/v1/providers/{id}/verify": {
      "post": {
        "tags": ["Providers"],
        "summary": "Verify provider connection",
        "description": "Test provider connectivity with a minimal LLM call.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "model": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Verification result", "content": { "application/json": { "schema": { "type": "object", "properties": { "valid": { "type": "boolean" }, "error": { "type": "string" } } } } } } }
      }
    },
    "/v1/skills": {
      "get": {
        "tags": ["Skills"],
        "summary": "List all skills",
        "responses": { "200": { "description": "Skill list" } }
      },
      "post": {
        "tags": ["Skills"],
        "summary": "Upload skill",
        "requestBody": { "content": { "multipart/form-data": {} } },
        "responses": { "201": { "description": "Skill uploaded" } }
      }
    },
    "/v1/skills/{id}": {
      "get": {
        "tags": ["Skills"],
        "summary": "Get skill",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Skill details" } }
      },
      "put": {
        "tags": ["Skills"],
        "summary": "Update skill",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Skill updated" } }
      },
      "delete": {
        "tags": ["Skills"],
        "summary": "Delete skill",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Skill deleted" } }
      }
    },
...
    "/v1/media/upload": {
      "post": {
        "tags": ["Media"],
        "summary": "Upload media file",
        "requestBody": { "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary" } } } } } },
        "responses": { "200": { "description": "Upload result with temp path and MIME type" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Gateway token or API key"
      }
    },
    "schemas": {
...
      "ProviderInput": {
        "type": "object",
        "required": ["name", "provider_type"],
        "properties": {
          "name": { "type": "string", "description": "Unique provider slug" },
          "display_name": { "type": "string" },
          "provider_type": { "type": "string", "enum": ["anthropic_native", "openai_compat", "gemini_native", "groq", "deepseek", "mistral", "xai", "ollama"] },
          "api_base": { "type": "string" },
          "api_key": { "type": "string" },
          "enabled": { "type": "boolean", "default": true }
        }
      },
...
    },
...
  }
}